backend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
black==22.3.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cssselect2==0.5.0
defusedxml==0.7.1
Django==4.0.3
django-appconf==1.0.5
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-haystack==3.1.1
django-imagekit==4.1.0
django-json-ld==0.0.5
django-qr-code==3.0.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.32.0
html5lib==1.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
pilkit==2.0
Pillow==9.1.0
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
PyJWT==2.3.0
pyphen==0.12.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.1
segno==1.4.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
weasyprint==54.3
webencodings==0.5.1
Whoosh==2.7.4
zopfli==0.2.1
